+++ src/xg.cpp
@@ -2705,11 +2705,11 @@ map<string, vector<size_t> > XG::distance_in_paths(int64_t id1, bool is_rev1, si
     return dist;
 }
 
-int XG::min_distance_in_paths(int64_t id1, bool is_rev1, size_t offset1,
-                              int64_t id2, bool is_rev2, size_t offset2) const {
+int64_t XG::min_distance_in_paths(int64_t id1, bool is_rev1, size_t offset1,
+                                  int64_t id2, bool is_rev2, size_t offset2) const {
     auto dist = distance_in_paths(id1, is_rev1, offset1,
                                   id2, is_rev2, offset2);
-    int min_dist = std::numeric_limits<int>::max();
+    int64_t min_dist = std::numeric_limits<int64_t>::max();
     for (auto& c : dist) {
         for (auto& o : c.second) {
             if (abs(o) < abs(min_dist)) {
